Tentativa 4

+++ constants.c
@@ -1,4 +1,4 @@
 const int SCREEN_WIDTH = 800;
 const int SCREEN_HEIGHT = 600;
 const int FPS = 30;
-const float tick_size = 1/(float)FPS;
+const float tick_size = 1/FPS;
